Changed all BigDecimal types to Double or double.

+++ src/pfnguyen/statlogic/algorithms/SortAlgorithms.java
@@ -1,6 +1,5 @@
 package pfnguyen.statlogic.algorithms;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 
 public class SortAlgorithms {
@@ -11,13 +10,13 @@ public class SortAlgorithms {
     /**
      * Inefficient sort for large lists
      */
-    public static void BubbleSort(ArrayList<BigDecimal> values) {
+    public static void BubbleSort(ArrayList<Double> values) {
         for (int i = 0; i < values.size(); i++) {
             for (int j = 0; j < values.size(); j++) {
 
                 if (j != values.size() - 1)
                     if (values.get(j).compareTo(values.get(j + 1)) > 0) {
-                        BigDecimal temp = values.get(j);
+                        Double temp = values.get(j);
                         values.set(j, values.get(j + 1));
                         values.set(j + 1, temp);
                     }
